multi chunk support and World manager implemented

+++ include/color.hpp
@@ -9,17 +9,17 @@
 
 // color type int uint32_t 0x ff(alpha) ff(blue) ff(green) ff(blue)
 
-const uint32_t colors[(int)Type::End] = {
+const uint32_t colors[(uint8_t)Type::End] = {
   0xffffffff, /* Air */
   0xff0fa5db, /* Sand */
   0xff0f0f0f , /* Stone */
   0xffb8291f, /* Water */
 };
 
-static_assert((int)Type::Air   == 0, "You need to adjust the colors");
-static_assert((int)Type::Sand  == 1, "You need to adjust the colors");
-static_assert((int)Type::Stone == 2, "You need to adjust the colors");
-static_assert((int)Type::Water == 3, "You need to adjust the colors");
+static_assert((uint8_t)Type::Air   == 0, "You need to adjust the colors");
+static_assert((uint8_t)Type::Sand  == 1, "You need to adjust the colors");
+static_assert((uint8_t)Type::Stone == 2, "You need to adjust the colors");
+static_assert((uint8_t)Type::Water == 3, "You need to adjust the colors");
 
 inline uint32_t GetColor(Type t){
   #ifdef DEBUG
@@ -28,7 +28,7 @@ inline uint32_t GetColor(Type t){
     return 0xffeb34cf;
   }
   #endif
-  return colors[(int)t];
+  return colors[(uint8_t)t];
 }
 
 #endif // COLOR_H_INCLUDED
